Trait orml_benchmarking::Benchmarking

pub trait Benchmarking {
    fn benchmarks(extra: bool) -> Vec<BenchmarkMetadata, Global> ;
    fn run_benchmark(
        name: &[u8],
        selected_components: &[(BenchmarkParameter, u32)],
        whitelist: &[TrackedStorageKey],
        verify: bool,
        internal_repeats: u32
    ) -> Result<Vec<BenchmarkResult, Global>, BenchmarkError>;
}

The pallet benchmarking trait.

Required Methods

fn benchmarks(extra: bool) -> Vec<BenchmarkMetadata, Global>

Get the benchmarks available for this pallet. Generally there is one benchmark per extrinsic, so these are sometimes just called “extrinsics”.

Parameters

extra: Also return benchmarks marked “extra” which would otherwise not be needed for weight calculation.

fn run_benchmark(
name: &[u8],
selected_components: &[(BenchmarkParameter, u32)],
whitelist: &[TrackedStorageKey],
verify: bool,
internal_repeats: u32
) -> Result<Vec<BenchmarkResult, Global>, BenchmarkError>

Run the benchmarks for this pallet.